sidebarLink folder rename, fix tests
